+++ packages/pds/tests/proxied/read-after-write.test.ts
@@ -9,6 +9,7 @@ import {
   AtpAgent,
 } from '@atproto/api'
 import { RecordRef, SeedClient, TestNetwork } from '@atproto/dev-env'
+import type { DidString } from '@atproto/syntax'
 import { app } from '../../src/lexicons/index.js'
 import basicSeed from '../seeds/basic.js'
 
@@ -17,8 +18,8 @@ describe('proxy read after write', () => {
   let agent: AtpAgent
   let sc: SeedClient
 
-  let alice: string
-  let carol: string
+  let alice: DidString
+  let carol: DidString
 
   beforeAll(async () => {
     network = await TestNetwork.create({